+++ lib/hop_tasks.dart
@@ -18,6 +18,7 @@ import 'package:hop/hop.dart';
 
 part 'src/hop_tasks/bench_task.dart';
 part 'src/hop_tasks/copy_directory.dart';
+part 'src/hop_tasks/dartanalyzer.dart';
 part 'src/hop_tasks/dart_analyzer.dart';
 part 'src/hop_tasks/dart2js.dart';
 part 'src/hop_tasks/dartdoc.dart';