+++ test/hop/integration_tests.dart
@@ -45,10 +45,6 @@ class IntegrationTests {
     // assuming `dart` is in system path
     env['PATH'] = Platform.environment['PATH'];
 
-    final options = new ProcessOptions()
-      ..workingDirectory = new Directory.current().path
-      ..environment = env;
-
-    return Process.run('dart', list, options);
+    return Process.run('dart', list, environment: env);
   }
 }