+++ test/hop_tasks/_hop_tasks.dart
@@ -15,6 +15,7 @@ import '../hop/_hop.dart';
 
 part 'process_tests.dart';
 part 'git_tests.dart';
+part 'dartanalyzer_tests.dart';
 part 'dart_analyzer_tests.dart';
 part 'copy_directory_tests.dart';
 part 'compiler_tests.dart';
@@ -25,6 +26,7 @@ void main() {
     GitTests.register();
     DartAnalyzerTests.register();
     CopyDirectoryTests.register();
+    Dart_AnalyzerTests.register();
     CompilerTests.register();
   });
 }